add NodeTraversal with `id` and `label`
