+++ traversals/src/main/scala/io/shiftleft/overflowdb/traversals/Traversal.scala
@@ -1,20 +1,17 @@
 package io.shiftleft.overflowdb.traversals
 
-import io.shiftleft.overflowdb.{NodeRef, OdbGraph}
-import org.apache.tinkerpop.gremlin.structure.Vertex
-
 import scala.collection.immutable.{ArraySeq, IndexedSeq}
 import scala.collection.{Iterable, IterableFactory, IterableFactoryDefaults, IterableOnce, IterableOps, Iterator, mutable}
-import scala.jdk.CollectionConverters._
 
 class Traversal[+A](elements: IterableOnce[A]) extends Iterable[A]
   with IterableOps[A, Traversal, Traversal[A]]
   with IterableFactoryDefaults[A, Traversal] {
+
+  def l: IndexedSeq[A] = elements.iterator.to(ArraySeq.untagged)
+
   override def className = "Traversal"
   override def iterableFactory: IterableFactory[Traversal] = Traversal
   override def iterator: Iterator[A] = elements.iterator
-
-  def l: IndexedSeq[A] = elements.iterator.to(ArraySeq.untagged)
 }
 
 object Traversal extends IterableFactory[Traversal] {
@@ -25,13 +22,4 @@ object Traversal extends IterableFactory[Traversal] {
   def from[A](source: IterableOnce[A]): Traversal[A] = new Traversal(Iterator.from(source))
 }
 
-abstract class TraversalSource(graph: OdbGraph) {
-  def all: Traversal[Vertex] = new Traversal(graph.vertices().asScala)
-
-  protected def nodesByLabel(label: String): Traversal[NodeRef[_]] =
-    new Traversal(graph.nodesByLabel(label).asScala)
-
-  protected def nodesByLabelTyped[A <: NodeRef[_]](label: String): Traversal[A] =
-    nodesByLabel(label).map(_.asInstanceOf[A])
-}
