+++ traversals/src/main/scala/io/shiftleft/overflowdb/traversals/package.scala
@@ -1,0 +1,6 @@
+package io.shiftleft.overflowdb
+
+package object traversals {
+  implicit def traversalToNodeTraversal[A <: NodeRef[_]](traversal: Traversal[A]): NodeTraversal[A] =
+    new NodeTraversal[A](traversal)
+}
