+++ traversals/src/test/scala/io/shiftleft/overflowdb/traversals/testdomains/gratefuldead/package.scala
@@ -4,10 +4,10 @@ import io.shiftleft.overflowdb.traversals.Traversal
 
 package object gratefuldead {
 
-  implicit def toArtistTraversal(trav: Traversal[Artist]): ArtistTraversal =
+  implicit def traversalToArtistTraversal(trav: Traversal[Artist]): ArtistTraversal =
     new ArtistTraversal(trav)
   
-  implicit def toSongTraversal(trav: Traversal[Song]): SongTraversal =
+  implicit def traversalToSongTraversal(trav: Traversal[Song]): SongTraversal =
     new SongTraversal(trav)
 
 }
